+++ ReleaseNotes.md
@@ -7,6 +7,9 @@
 ### Bug fixes
 - cta/CTA#1013 reportType uninitialized
 - cta/CTA#1035 log configuration options on frontend startup
+- cta/CTA#1042 Do not iterate over retrieve queues when holding global scheduler lock
+  - Repacks on a disabled tape must now use a mount policy whose name starts with repack
+  - There is no longer an empty mount when a disabled/broken tape queue is full of deleted requests
 
 # v4.3-1